+++ src/lib/merge-variables.js
